refactor: create jstz_utils
