+++ crates/jstzd/src/task/utils.rs
@@ -1,4 +1,5 @@
 use anyhow::{anyhow, Result};
+use jstz_utils::poll;
 use serde_json::Value;
 
 pub async fn retry<'a, F>(
@@ -19,24 +20,6 @@ where
     })
     .await
     .unwrap_or(false)
-}
-
-pub async fn poll<'a, F, T>(
-    max_attempts: u16,
-    interval_ms: u64,
-    f: impl Fn() -> F,
-) -> Option<T>
-where
-    F: std::future::Future<Output = Option<T>> + Send + 'a,
-{
-    let duration = tokio::time::Duration::from_millis(interval_ms);
-    for _ in 0..max_attempts {
-        tokio::time::sleep(duration).await;
-        if let Some(v) = f().await {
-            return Some(v);
-        }
-    }
-    None
 }
 
 pub async fn get_block_level(rpc_endpoint: &str) -> Result<i64> {
@@ -82,42 +65,6 @@ mod tests {
         let locked = Arc::new(Mutex::new(1));
         assert!(
             !super::retry(2, 1, || async { check(locked.clone(), true).await }).await
-        );
-    }
-
-    #[tokio::test]
-    async fn poll() {
-        async fn check(locked: Arc<Mutex<i32>>, result: bool) -> Option<bool> {
-            let mut v = locked.lock().await;
-            if *v == 5 {
-                return Some(result);
-            }
-            *v += 1;
-            None
-        }
-
-        // poll till the end and get a positive result
-        let locked = Arc::new(Mutex::new(1));
-        assert!(
-            super::poll(5, 1, || async { check(locked.clone(), true).await })
-                .await
-                .unwrap()
-        );
-
-        // poll till the end and get a negative result
-        let locked = Arc::new(Mutex::new(1));
-        assert!(
-            !super::poll(5, 1, || async { check(locked.clone(), false).await })
-                .await
-                .unwrap()
-        );
-
-        // not waiting long enough
-        let locked = Arc::new(Mutex::new(1));
-        assert!(
-            super::poll(2, 1, || async { check(locked.clone(), true).await })
-                .await
-                .is_none()
         );
     }
 }
